server/requirements.txt
-i https://pypi.org/simple
alembic==1.11.1 ; python_version >= '3.7'
aniso8601==9.0.1
blinker==1.6.2 ; python_version >= '3.7'
click==8.1.3 ; python_version >= '3.7'
commonmark==0.9.1
flask==2.3.2
flask-cors==3.0.10
flask-migrate==4.0.4
flask-restful==0.3.10
flask-sqlalchemy==3.0.3
flask-vite==0.1.6
greenlet==2.0.2 ; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==20.1.0
honcho==1.1.0
importlib-metadata==6.6.0
importlib-resources==5.12.0 ; python_version < '3.9'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
mako==1.2.4 ; python_version >= '3.7'
markupsafe==2.1.3 ; python_version >= '3.7'
psycopg2-binary==2.9.6
pygments==2.15.1 ; python_version >= '3.7'
python-dotenv==1.0.0
pytz==2023.3
rich==12.6.0 ; python_full_version >= '3.6.3' and python_full_version < '4.0.0'
setuptools==67.8.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlalchemy==2.0.16 ; python_version >= '3.7'
sqlalchemy-serializer==1.4.1
typing-extensions==4.6.3 ; python_version >= '3.7'
werkzeug==2.3.6 ; python_version >= '3.8'
zipp==3.15.0 ; python_version >= '3.7'
